selecttest/Selection-addRange.html

<!doctype html>
<title>Selection.addRange() tests</title>
<div id=log></div>
<script src=support/testharness.js></script>
<script src=support/testharnessreport.js></script>
<script src=common.js></script>
<script>
"use strict";

// TODO: Test addRange() on a non-empty Selection, once we have a usable spec
// for that.

function testAddRange(endpoints) {
	try {
		selection.removeAllRanges();
	} catch (e) {
		throw {message: "Exception thrown during removeAllRanges(): " + e};
	}
	try {
		var range = ownerDocument(endpoints[0]).createRange();
		range.setStart(endpoints[0], endpoints[1]);
		range.setEnd(endpoints[2], endpoints[3]);
	} catch (e) {
		throw {message: "Exception thrown during Range creation: " + e};
	}

	assert_equals(range.startContainer, endpoints[0],
		"Sanity check: the Range we created must have the desired startContainer");
	assert_equals(range.startOffset, endpoints[1],
		"Sanity check: the Range we created must have the desired startOffset");
	assert_equals(range.endContainer, endpoints[2],
		"Sanity check: the Range we created must have the desired endContainer");
	assert_equals(range.endOffset, endpoints[3],
		"Sanity check: the Range we created must have the desired endOffset");

	try {
		selection.addRange(range);
	} catch (e) {
		throw {message: "Exception thrown during addRange(): " + e};
	}

	assert_equals(range.startContainer, endpoints[0],
		"addRange() must not modify the startContainer of the Range it's given");
	assert_equals(range.startOffset, endpoints[1],
		"addRange() must not modify the startOffset of the Range it's given");
	assert_equals(range.endContainer, endpoints[2],
		"addRange() must not modify the endContainer of the Range it's given");
	assert_equals(range.endOffset, endpoints[3],
		"addRange() must not modify the endOffset of the Range it's given");

	assert_equals(selection.rangeCount, 1,
		"After removeAllRanges() and addRange(), rangeCount must be 1");
	assert_not_equals(selection.getRangeAt(0), null,
		"After addRange(), getRangeAt(0) must not return null");
	assert_equals(typeof selection.getRangeAt(0), "object",
		"After addRange(), getRangeAt(0) must return an object");

	assert_equals(selection.getRangeAt(0).startContainer, range.startContainer,
		"After addRange(), startContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).startOffset, range.startOffset,
		"After addRange(), startOffset of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endContainer, range.endContainer,
		"After addRange(), endContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endOffset, range.endOffset,
		"After addRange(), endOffset of the Range returned by getRangeAt(0) must match the added Range");

	assert_equals(selection.getRangeAt(0), range,
		"getRangeAt(0) must return the same object we added");

	// Let's not test many different modifications -- one should be enough.
	if (range.startContainer == paras[0].firstChild
	&& range.startOffset == 0
	&& range.endContainer == paras[0].firstChild
	&& range.endOffset == 2) {
		// Just in case . . .
		range.setStart(paras[0].firstChild, 1);
	} else {
		range.setStart(paras[0].firstChild, 0);
		range.setEnd(paras[0].firstChild, 2);
	}

	assert_equals(selection.getRangeAt(0).startContainer, range.startContainer,
		"After mutating the added Range, startContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).startOffset, range.startOffset,
		"After mutating the added Range, startOffset of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endContainer, range.endContainer,
		"After mutating the added Range, endContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endOffset, range.endOffset,
		"After mutating the added Range, endOffset of the Range returned by getRangeAt(0) must match the added Range");

	// Now test the other way too.
	if (selection.getRangeAt(0).startContainer == paras[0].firstChild
	&& selection.getRangeAt(0).startOffset == 4
	&& selection.getRangeAt(0).endContainer == paras[0].firstChild
	&& selection.getRangeAt(0).endOffset == 6) {
		selection.getRangeAt(0).setStart(paras[0].firstChild, 5);
	} else {
		selection.getRangeAt(0).setStart(paras[0].firstChild, 4);
		selection.getRangeAt(0).setStart(paras[0].firstChild, 6);
	}

	assert_equals(selection.getRangeAt(0).startContainer, range.startContainer,
		"After mutating the result of getRangeAt(0), startContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).startOffset, range.startOffset,
		"After mutating the result of getRangeAt(0), startOffset of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endContainer, range.endContainer,
		"After mutating the result of getRangeAt(0), endContainer of the Range returned by getRangeAt(0) must match the added Range");
	assert_equals(selection.getRangeAt(0).endOffset, range.endOffset,
		"After mutating the result of getRangeAt(0), endOffset of the Range returned by getRangeAt(0) must match the added Range");
}

var tests = [];
for (var i = 0; i < testRanges.length; i++) {
	tests.push(["Range " + i + " " + testRanges[i], eval(testRanges[i])]);
}

generate_tests(testAddRange, tests);

testDiv.style.display = "none";
</script>
